add-release-note: allow the workflow to be triggered by a GitHub App
The main purpose of this workflow is to be triggered via `/add relnote`
commands that are parsed by the GitForWindowsHelper GitHub App (see
https://github.com/git-for-windows/gfw-helper-github-app/commit/f26f1a2aa0a2
for details).

To that end, we better support the use case where a GitHub App triggers
this workflow.

That includes determining sensible Git authorship information, otherwise
the commit will fail. So let's do that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>