pacman-helper: simplify logic
In `push` and `push_missing_signatures`, we need to skip sources for
architectures other than x86_64, and MINGW packages for other
architectures than the current one.

The logic worked, but was convoluted, and unnecessarily verbose.

Let's fix this, and at the same time future-proof the code: We are about
to introduce support for packages targeting Windows/ARM64.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>