Add a script to downgrade Pacman packages
We just downgraded a couple of packages to accommodate for Git for
Windows sticking to OpenSSL v1.1.x while MSYS2 moved on to OpenSSL v3.x

Unfortunately, an upgrade of the `pacman` package undid this, so we had
to redo it.

Here is a script that was used to redo it. It takes as first parameter a
shell scriptlet that tests whether the issue that needed to be fixed was
actually fixed, followed by a list of packages to downgrade to the
previous version. This script was used to auto-generate
https://github.com/git-for-windows/git-sdk-64/commit/4c5ca7dbd83, by
calling

	/usr/src/build-extra/downgrade-pacman-package.sh \
		 'bsdtar -h || test $? != 127' \
		 bsdcpio bsdtar libarchive libserf libssh2 rsync

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>