installer: accommodate for the new location of `ca-bundle.crt`
Git for Windows' installer tries to be helpful to third-party software
that want to use the same certificate bundle as Git for Windows. To that
end, the location of the certificate bundle is written to the Git config
when using the OpenSSL is configured as SSL backend, even if Git for
Windows technically does not need it to be configured.

But the location changed in MSYS2, from `$MINGW_PREFIX/ssl/certs/` to
`$MINGW_PREFIX/etc/ssl/certs/`.

To accommodate for that location, we have to teach the installer about
it. And to make the code robust (i.e. to allow for running even with an
older version of the `mingw-w64-ca-certificates` package to be
installed), we look in the new location first, and fall back to the old
location.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>