mingit(BusyBox): do exclude Bash
The entire point of the BusyBox variant of MinGit is to avoid having to
use MSYS2 resources unnecessarily.

But `sh.exe` slipped through the net, by virtue of being listed in the
`sh` dependency (provided by `bash`) that comes in via openssh ->
heimdal -> heimdal-libs -> libedit -> sh.

This was pointed out in
https://github.com/git-for-windows/git/issues/3285#issuecomment-1396543785

Let's exclude it specifically.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>