build-installers: include the (to-be-created) .sparse/makepkg-git definition
The "Git Artifacts" Azure Pipeline is currently used to build the
installer/Portable Git/etc for official Git for Windows releases.

In the upcoming move to do the same in GitHub workflows in
https://github.com/git-for-windows/git-for-windows-automation, we want
to move away from building the i686 flavor of mingw-w64-git in the
x86_64 flavor of Git for Windows' SDK, and instead use the i686 flavor
of the SDK.

To this end, we will need to extend the i686 flavor of the
`build-installers` SDK artifact to be able to build that package.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>