please.sh: drop the obsolete `sync` command
Originally designed to keep clones of `git-sdk-32` and `git-sdk-64` in a
private VM up to date, over the course of the years this command became
more complex and also the driver of a CI job, which was then split into
two, ported from an Azure Pipeline to a GitHub workflow, and eventually
copy/edited into stand-alone GitHub workflows implemented mainly in
PowerShell, with minimal dependencies, in `git-sdk-32`, `git-sdk-64` and
`git-sdk-arm64`.

The subcommand itself is now superseded by these workflows.

This concludes https://github.com/git-for-windows/git/issues/4234

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>