gnupg-with-gpgkey: allow running on Ubuntu
In Git for Windows, we explicitly forbit Git from passing the
`--status-fd=2` option to GPG. But on Ubuntu, that option is passed, and
it is rather crucial because Git reads the status from `stderr` and is a
bit miffed if it does not get any status update there. The symptom is:

	error: gpg failed to sign the data
	error: unable to sign the tag

Let's prepare the script to allow for that option to be passed (and also
allow for that option _not_ to be passed).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>