pacman-helper quick_add: do handle symbolic links
When `MSYS` is set to `winsymlinks:nativestrict`, the MSYS2 runtime will
generate real Win32 symbolic links. This is something the `quick_add`
command did not expect, thinking that the `ln -s` command would generate
copies (which is the default of the MSYS2 runtime, believe it or not)
that can be overwritten. But it simply does not work because a symbolic
link cannot be overwritten with its target, not without removing the
link first.

This causes a problem since a recent `setup-git-for-windows-sdk` version
updated to a newer `@actions/toolkit` version that blatantly configures
the MSYS2 runtime to the `nativestrict` mode _for the entire remainder
of the workflow job_. For the full sadness of the details, see
https://github.com/actions/toolkit/pull/1291/commits/2867e318d4d0de
(yes, it indeed has no adequate commit message).

Let's work around this issue by removing symbolic links, if any, before
copying the databases or their signatures to the suffix-less version.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>