Release Notes: announce that 32-bit support is about to be deprecated
Since Git for Windows relies on the MSYS2 runtime for POSIX emulation as
required e.g. by the Bash (which is used by Git to execute aliases,
hooks, and the `git submodule` command), and since the MSYS2 project
already dropped support for the i686 variant of the MSYS2 runtime
(following Cygwin's lead), Git for Windows won't be able to provide full
support of 32-bit installers going forward.

Let's warn the users about this.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>