please.sh mention: rip out obsolete logic and hand off to add-release-note
When generating the release notes of v2.40.0-rc0, I noticed in
https://github.com/git-for-windows/git/pull/4316#issuecomment-1446624712
that a section was added unnecessarily.

The reason is that a crucial function was removed in
https://github.com/git-for-windows/build-extra/pull/457.

The plan was to retire the entire release process based on `please.sh`
and Azure Pipelines, and redo everything in git-for-windows-automation
via a `tag-git` workflow, but that workflow simply is not ready yet.

In preparation for those improved workflows, I already reimplemented the
`sed` logic of `please.sh mention [...]` with a more straight-forward
node.js script.

Let's just use that script in `please.sh mention` before phasing out the
functionality from `please.sh`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>