add-release-notes: allow for -rc<N> versions
We recently taught `add-release-note.js` to understand Git for Windows'
patch-level versions (e.g. "v2.39.0(2)").

This commit does the same for so-called "release candidates" (upstream
Git likes to call these release candidates, even if none of them
_actually_ are candidates for releasing). These versions end in `-rc<N>`
and while we never merge the updates to `ReleaseNotes.md` back into the
`main` branch, we do want to include release notes in the corresponding
Git for Windows installer that do mention that version.

This commit is best viewed with `--color-words="[A-Za-z0-9_]+|."`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>