add-release-note: avoid letting `Git` "supersede" `Git Credential Manager`
The current logic trying to prevent double entries for new component
versions was a bit loose and only checked the current component name as
a _prefix_.

But the _prefix_ `Git` matches not only the component name `Git`, but
also the `Git Credential Manager` one.

Let's be more careful to match the entire component name.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>