please.sh build-mingw-w64-git: make a smarter source package, faster
A while ago, mingw-w64-git source packages became so large that we could
no longer upload them to Azure Blobs. As a work-around, we started
re-packing the bare repository that is included in the source package to
fulfill GPL requirements.

However, the source packages are not only _still_ quite large, it also
takes a _loooong_ time to repack. Something like 4-5 minutes.

A better idea is to include only the relevant revision in the source
package. We now do this, by generating a `.tar.gz` archive of said
revision and then using a `PKGBUILD` file that refers to that archive
instead of a full clone.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>