please.sh build-mingw-w64-git: avoid updating pkgver unnecessarily
While the logic to source `PKGBUILD.$tag` and then calling the `pkgver`
function worked in a local test, in a GitHub workflow run it failed
(https://github.com/git-for-windows/git-for-windows-automation/actions/runs/4322469695/jobs/7544925582#step:18:3425):

	bash: line 1: pkgver: command not found

I introduced this call originally because my plan was to build the
Pacman package using the actual source package.

However, that would have modified the `mingw-w64-git/PKGBUILD` because
we commit and push this file as part of Git for Windows' release
process, and I do not actually want to commit the source archive but
stay with the process where we clone the git-for-windows/git repository.
Therefore, I switched the order back to building the binary package
first and only then building the source package (and not bleeding the
"build from source archive" change into the final `PKGBUILD` update).

Therefore, this call is no longer necessary and I can save me some time
by not investigating why it failed in CI when it worked locally.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>