pacman-helper: add a command to break a lease
To avoid interference between concurrent Pacman packages uploads, Git
for Windows' helper scripts first acquire a lock ("lease" in Azure
speak). Ideally, this lease would time out after something like 30
minutes, or even an hour, which would be a quite conservative maximum
estimate how long a successful upload could possibly take.

However, we cannot choose such a time-out because the maximum
non-infinite lease you can acquire for Azure Blobs is 60 seconds, see:
https://learn.microsoft.com/en-us/rest/api/storageservices/lease-blob.

Therefore, we _must_ acquire an infinite lease.

The downside is that we sometimes need to break that lease manually,
e.g. when a workflow run failed after acquiring such a lease (and hence
also failing to release it).

Previously, there was no tooling to do so, the only way to do it was to
direct a web browser to the Azure Portal, navigate to the `x86-64` Azure
Blob container and manually clicking the "Break lease" menu item.

This obviously adds to the stress when things go wrong (and the lease
only ever has to be broken when things go wrong). So, following the
mantra "deployments should be uneventful", we need to put tooling in
place that makes this operation much more of a no-brainer.

With this commit, there is now a scripted way for which we will add a
new GitHub workflow in git-for-windows/git-for-windows-automation, and
probably also teach the GitForWindowsHelper GitHub App a new trick
(`/break pacman-upload-lock` maybe?) to trigger that workflow.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>