please.sh(coverity): fix sanity check
The `submit-build-to-coverity` command has a sanity check to validate
that it is really working on a Git checkout.

However, this fails to work with shallow clones because it refers to the
initial commit of git.git.

Besides, it was missing a crucial `-C "$git_src_dir"`, i.e. it was
broken to begin with.

Let's even allow building from source code extracted from a tar file,
simply by relying on the presence of a well-known and pretty distinctly
Git-only file.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>