README.md: report Issues via our main Git-for-Windows git fork
Issue reporting is switched off for the build-extra repository. The
Git-for-Windows organisation has many potentially relevant repos for
reporting issues. Let's direct queries to the primary repo, rather than
say the "SDK" repositories.

While here, clarify that the build-extra repo is a core part of the SDK
experience, rather than being the SDK in it's totality.

Fixes: https://github.com/git-for-windows/git/discussions/4355 step 1.

Signed-off-by: Philip Oakley <philipoakley@iee.email>