installer: allow silent install of 32-bit installer on ARM64
When Installing 32-bit Git for Windows we show a Warning that we're deprecating
32-bit support, but there are two Versions of the 32-bit warning. The Version
shown on 32-Bit Windows allows a silent install, while the Version on 64-Bit
Windows does not. This is fine on AMD64, but for Windows 10 on ARM64
we don't have a good alternative offer. So show them the same message as we show
on i686.

Signed-off-by: Matthias Aßhauer <mha1993@live.de>