git-extra: install an extra linter to prevent incorrect MSYS packages
A recent change in `libgpg-error` caused the i686 variant to be built
using the incorrect filename `cyggpg-error-0.dll`. The reason is that we
tried to align better with MSYS2, but they already dropped support for
i686 long ago and therefore `git-sdk-32`'s autoconf is ill-prepared and
did not build the intended `msys-gpg-error-0.dll` file.

To prevent similar breakages in the future _before_ deploying the broken
packages, let's add a custom linter to catch precisely this type of
issue.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>