make-file-list: include libcurl alternate when appropriate
As of https://github.com/git-for-windows/MINGW-packages/pull/76, Git for
Windows wants to ship with _two_ libcurl variants: one that includes the
Secure Channel backend to talk https://, and another one that includes
the OpenSSL backend instead.

Therefore, the `make-file-list` script that is used by Git for Windows'
automation to list the files required for the various release assets
needs to be adjusted, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>