installer: switch the libcurl default to Secure Channel
In line with https://github.com/git-for-windows/MINGW-packages/pull/76,
the default SSL/TLS backend is now the Secure Channel one.

Let's reflect that in the installer.

This makes Git for Windows' maintenance easier (see
https://github.com/git-for-windows/git/issues/4350 for a fuller
explanation). But it has further benefits: the Secure Channel backend
is the "more native" choice on Windows, as it integrates better with
Windows. For example, when trusting a host with a self-signed
certificate in a web browser, that trust extends to Secure Channel. With
the OpenSSL backend, users had to jump through manual hoops to get `git
clone <url>` to trust a host with a self-signed certificate.

Note: The installer will _still_ pick up previous choices when upgrading
from an earlier Git for Windows version. Only fresh installs without any
current Git for Windows installations will receive the new default.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>