installer: adjust pre-flight checklist for the libcurl split
At the moment, the command `git -c http.sslBackend=openssl <...>` will
fail because there is not yet any new mingw-w64-git package version that
supports dynamically loading various libcurl variants depending on the
configured backend choice.

This is not a terrible thing because Git for Windows v2.41.0-rc0 is
imminent (due either later today, or at the latest tomorrow), and that
will address the issue.

So let's adjust the checks accordingly, to be lenient in case they are
run with Git for Windows v2.40.1 (i.e. the still-current newest package
version).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>