apply-from-lore: support patch series iterations better
When a, say, v2 of a patch series is posted, we will want to get the
patches from that v2, not v1, not v3.

Typically, these are identified by the `v2` part of `[PATCH v2 00/15]`
in the cover letter's subject line, and that `v2` part is repeated
throughout the individual patches.

Let's make sure that we match the precise `PATCH v<N>` prefix so that we
find the intended patches.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>