installer: offer a workaround for mandatory ASLR
ASLR (Address Space Layout Randomization) is an important tool to secure
software, by avoiding predictable addresses from which malicious shell
code could be executed in attacks exploiting vulnerabilities.

Unfortunately, the way the MSYS2 runtime works is incompatible with ASLR
and would not allow processes to be spawned via the emulated `fork()`
syscall. The MSYS2 runtime, though, is the work horse behind Git Bash
and all the other executables in `/usr/bin/`.

With this change, the installer detects when mandatory ASLR is enabled,
and in that case offers the user to add security exceptions so that the
executables in the "usr/bin" directory and Git Bash function correctly.

The main logic is implemented in a PowerShell script that can be run
separately from the installer, too.

Seeing as it is a security concern to disable mandatory ASLR in such a
big sweep, let's warn the users loudly (and with red "WARNING" labels)
about it.

Note: The documentation at
https://learn.microsoft.com/en-us/windows/security/threat-protection/overview-of-threat-mitigations-in-windows-10#converting-an-emet-xml-settings-file-into-windows-10-mitigation-policies
suggests to use the invocation

	Set-ProcessMitigation [...] -Disable MandatoryASLR

However, as the documentation at
https://learn.microsoft.com/en-us/powershell/module/processmitigations/set-processmitigation
clearly documents (and, `Get-Help Set-ProcessMitigation`, too), the
`MandatoryASLR` value is not applicable, but the `ForceRelocate` one is.
The example even mentions `ForceRelocateImages`, which is what we use,
then.

Also note: There is apparently no official documentation indicating how
to detect when mandatory ASLR is enabled. The best documentation found
in the process of developing this here patch is at
https://learn.microsoft.com/en-us/microsoft-365/security/defender-endpoint/troubleshoot-exploit-protection-mitigations
and suggests looking at `HKLM:\SOFTWARE\Microsoft\Windows
NT\CurrentVersion\Image File Execution Options`, which is what this
patch does.

Signed-off-by: Andrew Riseborough <AndyBorah38@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>