installer: infer default branch name from system config, if set
When a user configures the default branch name manually, i.e. edits the
system config themselves, the Git for Windows installer currently resets
it to whatever was specified when running the installer previously.

This is inconsistent with the way other defaults are inferred from the
system config, say, `core.symlinks`.

Let's make it consistent by interpreting the current value of
`init.defaultBranch` (if configured).

This addresses https://github.com/git-for-windows/git/issues/4525.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>