install.iss: when silent, do not prompt for process closure
This block that is being edited is already within an "if WizardSilent()"
block, so we should not be prompting via a message box.

Instead, fail with a log message indicating the problem.

The diff looks a bit nasty since we are removing a retry loop at the
same time. Hide whitespace changes to make it more reasonable.

Signed-off-by: Derrick Stolee <derrickstolee@github.com>