installer: fix the broken "Enable Pseudo Console Support" option
In 3cc29a613 (installer: explicitly turn off pseudo console support,
when asked, 2023-05-16), I tried to explicitly turn off pseudo console
support when the checkbox is unchecked.

However, the change contains a bug: While the indentation clearly
relates the intention to have an `if pcon_enabled ... else ...` type of
flow, the code block before the `else` was not terminated, and therefore
it was essentially an `if pcon_enabled ...` type of flow, i.e. it did
_nothing_ when the checkbox was unchecked.

Even worse: if the checkbox _was_ checked, the added `else` block would
now be triggered if the `git-bash.config` was _written correctly_, but
would now incorrectly _overwrite_ the contents (which would read
"MSYS=enable_pcon" as intended) with "MSYS=disable_pcon".

In essence, the bug where the block before the `else` was not correctly
terminated (and the block after the `else` was not correctly opened)
leads to _the opposite_ of the intended logic:

- If the "Enable Pseudo Console Support" box is checked, we would
  eventually write the config file such that Pseudo Console support is
  _disabled_!

- If the box is left unchecked, we would not write `git-bash.config` at
  all, going with the default of the MSYS2 runtime, which is _to enable_
  Pseudo Console support!

Let's correct this unintentional behavior by fixing the bug.

This fixes https://github.com/git-for-windows/git/issues/4571.

Reported-by: Erik Falor <ewfalor@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>