installer checklist: suggest to run the script
To verify that Git for Windows installer works as intended, there is a
"pre-flight check list" in `checklist.txt`.

A few of the steps listed in that file have been performed manually with
every Git for Windows release and pre-release for years, until I got
tired of doing them manually (and I missed a few every once in a while),
which is when I started putting as much as I could into the
`run-checklist.sh` script.

Instead of duplicating the information in `checklist.txt` that is
already contained in the script, let `checklist.txt` suggest to run the
script instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>