installer: avoid trying to copy a non-existent file
As of https://github.com/git-for-windows/git/pull/4252, the "dashed"
command `git-log.exe` is no longer provided by the Git for Windows
project.

However, our very own installer expected it still to be there, in case a
fall-back was needed when hard-linking the "dashed" built-in commands.
Which resulted in beautiful (non-fatal) error messages like:

	Line 3029: Creating copy "C:\Program Files (x86)\Git\tmp\git-wrapper.exe" failed.
	Line 3053: Deleting temporary "C:\Program Files (x86)\Git\tmp\git-wrapper.exe" failed.

Since the intention was always to use the `git-wrapper.exe` file for
that fall-back, let's use that file directly instead, and avoid copying
any file preemptively, whether it exists or not.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>