installer: adjust a misleading log message
When readers like myself see the word "failed", all kind of
neurotransmitters start doing work.

When the only issue at hand is that a check was not performed because
it needs to be run in non-admin mode and we happen to have been called
as admin and therefore have no way to perform this check, it is
incorrect to even say something _failed_. It was _skipped_, is all.

In preparation for adding a stringent check to the PR build that
verifies that working installers can be built reliably, let's tone down
that message.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>