Stop using p7zip, use 7-Zip instead
Back in the days, 7-Zip was only a Windows GUI program, and p7zip was
started as a fork to provide a CLI program that could also be compiled
and used on Linux. Since it was easier to port p7zip than 7-Zip to
MSYS2, we used the former to build the PortableGit, and even to squeeze
out more compression for the MinGit `.zip`.

Sadly, p7zip is languishing. The latest version of that project is from
2016: https://sourceforge.net/projects/p7zip/files/.

Happily, 7-Zip was modified in the meantime to allow building CLI
programs on Linux, and based on that support, I was able to port it to
MSYS2: https://github.com/msys2/MINGW-packages/pull/13894.

Yesterday, I updated it to match 7-Zip v23.01:
https://github.com/msys2/MINGW-packages/pull/18254, and this is as fine
an opportunity as any to start moving away from a stale p7zip.

This should also be a bit quicker, as we're moving from an MSYS2
`7za.exe` to a MINGW `7z.exe`.

Also, 7-Zip v23.01 improved the compression. Not by a whole lot
(PortableGit shrank by about 200kB, which is not much given that the
entire file is around 57MB to begin with), but hey, it's something.

Note: The change to `portable/release.sh` is a bit more complicated
because 1) the command-line got too long (MSYS2 side-steps the 32-kB
length limit by playing games with shared memory), and 2) the absolute
path needs to be provided as a _Windows_ path now, not as a pseudo-Unix
path.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>