ci: always show the `installer.log`
We do log the output into a file so that it can be inspected, but
currently the output is not shown unless the installation succeeded.

That's bad. We need the output _particularly_ when the installation
failed.

Let's just print it in a separate step, making sure that it is printed
whenever the installation fails.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>