ci: avoid bogus error
The idea is to force the `sdk-artifacts` job to be run when more than
200 lines are in the numstat of `please.sh`.

However, when the numstat of `please.sh` is empty, the check failed
because the empty string is not 0.

Abuse a quirk in Bash where `$(())` (i.e. the arithmetic evaluation of
the empty string) yields 0.

This avoids symptoms in the CI runs' logs of the form:

	/home/runner/work/_temp/17a8b050-ac43-4536-9818-63b75d561be8.sh: line 23: test: 200: unary operator expected

Incidentally, this bug ensured that the `sdk-artifacts` job was run
always, not only when necessary.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>