ci: avoid building the installer multiple times
Under certain circumstances, the installer would be built both in the
`build-artifacts` matrix as well as in the `sdk-artifacts` one.

That is a waste of time and resources, so let's just build the installer
in the latter matrix (which validates both 32-bit and 64-bit build, and
is therefore a bit more comprehensive than the former matrix in that
instance).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>