ci: validate the installer also in the `build-artifacts` job
We have this awkward situation where the installer can be built as part
of the `sdk-artifacts` matrix as well as of the `build-artifacts` one.

In the former case, we validated the installer, in the latter we did
not.

Now that we no longer build the installer in both, we must copy/edit the
validation steps into the latter matrix.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>