installer/portable: include libbrotlienc
As of brotli 1.1.0, it seems that the `brotli` executable depends on
that library. And since we include that executable, we need to include
that library, too.

An alternative would be to kick out that executable, as it is _probably_
not required for libcurl to work (which is the only user of brotli in
Git for Windows). But that would be fragile (what if the analysis is
incomplete and the executable _is_ required, in corner cases?) and the
library only weighs ~64 kilobyte anyway, so let's go the robust and easy
route.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>