build-installers: do include the 7-Zip DLL
In 537c34a3d (Stop using p7zip, use 7-Zip instead, 2023-08-25), I
switched Git for Windows to using the MINGW variant of 7-Zip instead of
the MSYS variant of p7zip.

However, there was a quite obvious mistake where I specified 7z.exe to
be included in the `build-installers` sdk-artifact _twice_, and 7z.dll
not at all.

The intention was to specify both, one time each, which is what this
here commit does.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>