installer: avoid cut-off text
As reported in https://github.com/git-for-windows/git/issues/4727,
currently the installer shows labels that are cut off. It is a bit
unclear what caused this: While I first suspected the new Windows font,
Aptos, to be the culprit, that hypothesis was rejected when another
contributor reported that the text appears not to be cut off for them
(https://github.com/git-for-windows/build-extra/pull/536#issuecomment-1847227487).

Let's address this by re-wrapping the text more conservatively, so that
the text appears intact in more setups.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>