Release Notes: Announce the end of Windows 7/8 support _after_ v2.44
Cygwin v3.5.0 (which Git for Windows uses in the form of an adjusted
MSYS2 runtime) drops Windows 7/8 support.

Git for Windows v2.44 comes a little too close after the release of
Cygwin v3.5.0 (and after a critical bug has been spotted immediately in
that version that would prevent `git.exe`'s exit code to be propagated
correctly to the calling MSYS2 Bash), therefore Git for Windows v2.44
will still come with MSYS2 runtime v3.4.*.

The good news is that we can now set a firm date when Git for Windows
will drop Windows 7/8 support: In the next version (which is likely
Git for Windows v2.45).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>