Release Notes: add back deprecation notices
I just noticed that those notices are not included in the -rc0 and -rc1
of v2.45.0, but let's add them back for the final v2.45.0 release
scheduled for this coming Monday, April 29th, 2024.

Please note that v2.45.0 will still support Windows 7/8 due to the move
to MSYS2 runtime v3.5.x being delayed (the MSYS2 project did not make
that switch yet, for similar reasons, and also because the quick
succession of the .0, .1, .2 and .3 releases raised doubts.).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>