portable: make sure that `package-versions.txt` is included
In https://github.com/git-for-windows/git/issues/4955 it was pointed out
that the i686 version ("32-bit") of Portable Git v2.45.1 is missing the
`package-versions.txt` file.

While it is unclear why that happened (the logs of the -- embargoed --
build do not reveal any clue about this issue), let's at least ensure
that it does not happen again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>