Fix user path reg query potentially getting truncated
The reg query command used returns a multi-line output that contains the path variable's content in its third line:
    Path    REG_EXPAND_SZ    [actual path variable]
The previous code just returned the last column of that line using awk, which fails to account
for spaces in the path, in which case only the part after the last space was printed.
Now cut is used to skip the first 12 spaces and print every after it, which returns the entire path variable's content unmodified.

Signed-off-by: Moritz Bender <35152647+Morilli@users.noreply.github.com>