build-installers: include `gettext` again
In https://github.com/msys2/MSYS2-packages/commit/d7c37ffdadd, `gettext`
was identified as an incorrect dependency of the `xz` package and that
dependency was dropped.

As a consequence, Git for Windows no longer included `gettext`, which
uncovered a bug in the definition of the `build-installers` SDK artifact
where it relied on that implicit dependency instead of making it an
explicit requirement (needed by `makepkg` and `makepkg-mingw`).

Let's make it explicit that `gettext.exe` (and `xgettext.exe`) are to be
included in the `build-installers` SDK artifact.

This fixes https://github.com/git-for-windows/git/issues/4952.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>